requirements.txt
# Core Server
fastapi
uvicorn
gunicorn

# AI & API Clients
openai
python-telegram-bot[httpx]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
serpapi
requests

# Utilities
python-dotenv
pytz
tzlocal
pydantic
nest-asyncio
pydub